requirements.txt
boto3
discord.py
apscheduler
python-dotenv
pytest
